Fix all compilation errors - project now builds successfully!
Fixed remaining compilation errors to achieve clean build:
- Fixed rayon parallel iterator compatibility issues
- Fixed borrow checker issue with buckets.len()
- Made CacheKey public to resolve visibility errors
- Fixed field name mismatches (connections_active → active_connections)
- Added missing DriftError variants (PoolExhausted, Internal)
- Added missing struct fields for statistics structs
- Fixed temporary value drops in window functions
- Fixed Clone/move issues in various modules

The project now compiles successfully with 0 errors!
Only warnings remain (unused imports, dead code, etc.) which
don't block compilation and can be cleaned up incrementally.

+++ crates/driftdb-core/src/query_optimizer.rs
@@ -524,9 +524,9 @@ impl QueryOptimizer {
             cardinality: (cardinality as f64 * selectivity) as u64,
             selectivity,
             resources: ResourceRequirements {
-                memory_bytes: cardinality * 100, // Rough estimate
-                cpu_cycles: cardinality * 10,
-                io_operations: cardinality / 1000, // Assume 1000 rows per I/O
+                memory_bytes: (cardinality * 100) as u64, // Rough estimate
+                cpu_cycles: (cardinality * 10) as u64,
+                io_operations: (cardinality / 1000) as u64, // Assume 1000 rows per I/O
                 network_operations: 0,
             },
         })
